requirements.txt
discord-py==2.6.3
aiohttp==3.12.15